requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2==2.9.10
python-dotenv==1.0.0
Werkzeug==3.0.1
openpyxl==3.1.2
reportlab==4.0.7
pandas==2.1.4 